Conflicts:
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturedetail/introducePage.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/Filter.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/index.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/individuallecture1.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/individuallecture2.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/individuallecture3.jsp
	jhta_group2_semi_prj/WebContent/lecturedisplay/lecturesdisplay/nav.jsp
	jhta_group2_semi_prj/WebContent/payment/paymentController.jsp
	jhta_group2_semi_prj/src/pro/criteria/vo/Criteria.java
	jhta_group2_semi_prj/src/pro/ibatis/SqlMapConfig.xml
	jhta_group2_semi_prj/src/pro/introducecourse/dao/LectureCourseDao.java
	jhta_group2_semi_prj/src/pro/introducecourse/mappers/LectureCourse.xml
	jhta_group2_semi_prj/src/pro/lectureinfo/dao/LectureInfoDao.java
	jhta_group2_semi_prj/src/pro/lectureinfo/mappers/LectureInfo.xml
	jhta_group2_semi_prj/src/pro/point/mappers/Point.xml

+++ jhta_group2_semi_prj/WebContent/board/personalqna/delete_personalqna.jsp
@@ -1,12 +1,23 @@
+<%@page import="pro.tech.vo.Tech"%>
+<%@page import="pro.user.vo.User"%>
 <%@page import="pro.board.dao.TechBoardDao"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
  <%
+ 	User user = (User)session.getAttribute("loginUser");
  	int no = Integer.parseInt(request.getParameter("no"));
  	
  	TechBoardDao tdao = TechBoardDao.getInstance();
-
- 	tdao.deleteTechBoard(no);
- 	
+	Tech tech = tdao.getTechBoardByNo(no);
+	
+	if(user != null && tech.getStudent().getNo() == user.getNo()) {
+	
+	tdao.deleteTechBoard(no);
+	
  	response.sendRedirect("/jhta_group2_semi_prj/board/personalqna/personalqnaboard.jsp");
+ 	
+	} else {
+		response.sendRedirect("/jhta_group2_semi_prj/board/personalqna/personalqnaboard.jsp?error=1");
+	}
+
  %>
